Break item into component
