+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/shared/right-panel/headless-treeview/index.ts
@@ -1,0 +1,3 @@
+export { TreeView } from "./TreeView";
+export { TreeItem } from "./TreeItem";
+export type { TreeItemData, TreeItemInstance } from "./types";
